LayoutTests/fast/css/webkit-filter-calc-expected.txt

Tests assigning a calculated value to '-webkit-filter' CSS property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testDiv.style['-webkit-filter'] is ""
testDiv.style['-webkit-filter'] = 'saturate(calc(10% * 2))'
PASS testDiv.style['-webkit-filter'] is "saturate(calc(20%))"
PASS window.getComputedStyle(testDiv).getPropertyValue('-webkit-filter') is "saturate(0.2)"
testDiv.style['-webkit-filter'] = 'brightness(calc(10% * 2))'
PASS testDiv.style['-webkit-filter'] is "brightness(calc(20%))"
PASS window.getComputedStyle(testDiv).getPropertyValue('-webkit-filter') is "brightness(0.2)"
PASS successfullyParsed is true

TEST COMPLETE
